Custom view
- Add Option to have custom view inside the tooltip
- Support animaton.
- Support arrow on custom view.
- Extract more classes outside of Tooltip.
