+++ app/src/main/res/layout/activity_main.xml
@@ -189,6 +189,12 @@
                 android:text="@string/click_me" />
 
             <com.google.android.material.button.MaterialButton
+                android:id="@+id/customViewTooltip"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="Custom View Tooltip" />
+
+            <com.google.android.material.button.MaterialButton
                 android:id="@+id/button2"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
