+++ xtooltip/src/main/java/it/swabbass/android/library/xtooltip/Gravity.kt
@@ -1,0 +1,5 @@
+package it.swabbass.android.library.xtooltip
+
+enum class Gravity {
+    LEFT, RIGHT, TOP, BOTTOM, CENTER
+}
